Removing SD web image and showing the avatar in the thread experience.
