+++ Shared/Protocols/ImageDisplayable.swift
@@ -12,7 +12,6 @@ import Parse
 protocol ImageDisplayable {
     var userObjectID: String? { get }
     var image: UIImage? { get }
-    var url: URL? { get }
     var textResult: NSTextCheckingResult? { get }
 }
