+++ Shared/UI/Views/MessageContentView/MessageContentView.swift
@@ -83,6 +83,7 @@ class MessageContentView: View {
 
     func configure(with message: Messageable) {
         self.message = message
+
         if message.isDeleted {
             self.textView.text = "DELETED"
         } else {
