+++ iOS/Flows/Conversation/Cells/MessageSubcell.swift
@@ -46,8 +46,9 @@ class MessageSubcell: UICollectionViewCell {
         }
     }
 
-    func configure(with message: Messageable) {
+    func configure(with message: Messageable, showAuthor: Bool) {
         self.content.configure(with: message)
+        self.content.state = showAuthor ? .expanded : .collapsed
         self.detailView.configure(with: message)
         self.layoutNow()
     }
