+++ iOS/Flows/Conversation/ConversationViewControllerDataSource+CellRegistration.swift
@@ -30,7 +30,7 @@ extension ConversationCollectionViewDataSource {
             let messageController = ChatClient.shared.messageController(cid: item.channelID,
                                                                         messageId: item.messageID)
             guard let message = messageController.message else { return }
-            cell.configure(with: message)
+            cell.configure(with: message, showAuthor: true)
         }
     }
